requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
Django==3.0.7
fake-useragent==0.1.11
google-auth==1.18.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
webdriver-manager==3.2.1
